Improve portability by bundling deps locally.
The hardcoded paths did not work on Ubuntu 17.10. Besides,
there are a number of other directories in the path to consider
not just these, so there are other ways the old method to fail.

The new behavior tries to be smarter by installing deps directly
into the dist folder in the first place, removing the need to copy
the files later.